Mark filldiff as finished after applied to remote. (#847)

+++ smart-metastore/src/main/java/org/smartdata/metastore/dao/FileDiffDao.java
@@ -91,7 +91,7 @@ public class FileDiffDao {
     simpleJdbcInsert.executeBatch(maps);
   }
 
-  public int update(int did, boolean newApplied) {
+  public int update(long did, boolean newApplied) {
     JdbcTemplate jdbcTemplate = new JdbcTemplate(dataSource);
     String sql = "update file_diff set applied = ? WHERE did = ?";
     return jdbcTemplate.update(sql, newApplied, did);
